Use provider compatible #callback_url impl
The implementation of #callback_url in omniauth-oauth2 strategy
changed in a way that is correct according to oauth spec, but breaks
many oauth providers.

This change adapts this strategy to the old behavior. That way
omniauth-oauth2 1.4 can be used. That is better, as version 1.3.1 seems
to be incompatible with Rails 5.